Avoid undefined behavior in MPL3115A2 altitude left-shift
Left-shifting a negative int32_t is undefined behavior in C.
Cast to uint32_t for the left shift, then back to int32_t for
the arithmetic right shift which is well-defined on GCC/ARM.

+++ drivers/sensor/mpl3115a2/mpl3115a2.c
@@ -155,7 +155,7 @@ static int sample_altitude(const struct device *dev)
 	if (raw & 0x800000) {
 		raw |= 0xFF000000;
 	}
-	get_data(dev)->altitude = ((raw << 8) >> 12) / 16.f;
+	get_data(dev)->altitude = ((int32_t)((uint32_t)raw << 8) >> 12) / 16.f;
 	get_data(dev)->temperature = (int8_t)buffer[3] + (buffer[4] >> 4) / 16.f;
 
 	return 0;
